remove BaseLinter
